+++ vm/src/vm/vm_core.rs
@@ -4,21 +4,24 @@ use crate::types::builtin_name::BuiltinName;
 #[cfg(feature = "extensive_hints")]
 use crate::types::program::HintRange;
 use crate::{
-    hint_processor::hint_processor_definition::HintProcessor,
+    hint_processor::{
+        builtin_hint_processor::blake2s_hash::blake2s_compress,
+        hint_processor_definition::HintProcessor,
+    },
     types::{
         errors::math_errors::MathError,
         exec_scope::ExecutionScopes,
         instruction::{
             is_call_instruction, ApUpdate, FpUpdate, Instruction, Opcode, PcUpdate, Res,
         },
-        relocatable::{MaybeRelocatable, Relocatable},
+        relocatable::{MaybeRelocatable, MaybeRelocatable::RelocatableValue, Relocatable},
     },
     vm::{
         context::run_context::RunContext,
         decoding::decoder::decode_instruction,
         errors::{
             exec_scope_errors::ExecScopeError, memory_errors::MemoryError,
-            vm_errors::VirtualMachineError,
+            memory_errors::MemoryError::AddressNotRelocatable, vm_errors::VirtualMachineError,
         },
         runners::builtin_runner::{
             BuiltinRunner, OutputBuiltinRunner, RangeCheckBuiltinRunner, SignatureBuiltinRunner,
@@ -438,8 +441,58 @@ impl VirtualMachine {
             .memory
             .mark_as_accessed(operands_addresses.op1_addr);
 
+        if instruction.opcode == Opcode::Blake2s {
+            self.handle_blake2s_instruction(&operands)?;
+        }
+
         self.update_registers(instruction, operands)?;
         self.current_step += 1;
+
+        Ok(())
+    }
+
+    /// Executes a Blake2s instruction.
+    /// Expects operands to be RelocatableValue and to point to segments of memory.
+    /// op0 is expected to point to a sequence of 9 u32 values (state and counter t0).
+    /// op1 is expected to point to a sequence of 16 u32 values (message).
+    /// dst is expected to point to a sequence of 8 cells each being either unitialised or
+    /// containing the Blake2s compression output at that index.
+    /// Deviation from the aforementioned expectations will result in an error.
+    /// The instruction will update the dst memory segment with the new state.
+    /// Note: the byte counter should count the number of message bytes processed so far including
+    /// the current portion of the message (i.e. it starts at 64, not 0).
+    fn handle_blake2s_instruction(
+        &mut self,
+        operands: &Operands,
+    ) -> Result<(), VirtualMachineError> {
+        let dst: Relocatable = match operands.dst {
+            RelocatableValue(relocatable) => relocatable,
+            _ => return Err(VirtualMachineError::Memory(AddressNotRelocatable)),
+        };
+
+        let state_and_t0: [u32; 9] = (self.get_u32_range(&operands.op0, 9)?)
+            .try_into()
+            .map_err(|_| VirtualMachineError::Blake2sInvalidOperand(0, 9))?;
+        let message = (self.get_u32_range(&operands.op1, 16)?)
+            .try_into()
+            .map_err(|_| VirtualMachineError::Blake2sInvalidOperand(1, 16))?;
+
+        let mut new_state = vec![];
+
+        new_state.extend(blake2s_compress(
+            state_and_t0[0..8].try_into().unwrap(),
+            &message,
+            state_and_t0[8],
+            0,
+            0,
+            0,
+        ));
+
+        for (i, &val) in new_state.iter().enumerate() {
+            self.segments
+                .memory
+                .insert_as_accessed((dst + i)?, MaybeRelocatable::Int(Felt252::from(val)))?;
+        }
 
         Ok(())
     }
@@ -924,6 +977,33 @@ impl VirtualMachine {
         size: usize,
     ) -> Result<Vec<Cow<Felt252>>, MemoryError> {
         self.segments.memory.get_integer_range(addr, size)
+    }
+
+    /// Gets n u32 values from memory starting from addr (n being size),
+    /// Verifies that addr is &RelocatableValue and that the values are u32.
+    pub fn get_u32_range(
+        &self,
+        addr: &MaybeRelocatable,
+        size: usize,
+    ) -> Result<Vec<u32>, MemoryError> {
+        let addr_relocatable: Relocatable = match addr {
+            &RelocatableValue(relocatable) => relocatable,
+            _ => return Err(MemoryError::AddressNotRelocatable),
+        };
+        let res_cow = self.get_integer_range(addr_relocatable, size)?;
+        let mut res = vec![];
+        for (i, x) in res_cow.iter().enumerate() {
+            let le_digits = x.clone().into_owned().to_le_digits();
+            if le_digits[0] >= (1 << 32)
+                || le_digits[1] != 0
+                || le_digits[2] != 0
+                || le_digits[3] != 0
+            {
+                return Err(MemoryError::ExpectedU32(Box::new((addr_relocatable + i)?)));
+            }
+            res.push(le_digits[0] as u32);
+        }
+        Ok(res)
     }
 
     pub fn get_range_check_builtin(